Change expectedCaloriesPerDay to Int

+++ src/main/java/com/client/calorieserver/domain/dto/CreateUserRequest.java
@@ -24,5 +24,5 @@ public class CreateUserRequest {
     //TODO custom validator to check if role names exist
     private Set<String> roles;
 
-    private Long expectedCaloriesPerDay;
+    private Integer expectedCaloriesPerDay;
 }
